+++ xdef/changelog.md
@@ -1,11 +1,13 @@
 # Version ${version}, release-date ${release.date}
 
-# Version 41.0.4, release-date 2021-12-31
+# Version 41.0.4, release-date 2022-01-01
 * added new type validation methods `country` and `countries`
 * `<xd:json>` model in X-definition changed to `<xd:xon>`
 * fixed bug in datetime formationg when mask of a xdatetime contains `SSS`
   and the value of milliseconds in datetime is zero
 * fixed bug NullpointerException in parsing of incorrect JSON data
+* corrected generation of XON from XComponent
+  (`org.xdef.component.XComponentUtil.toXon(...)`.
 
 # Version 41.0.3, release-date 2021-12-05
 * added tools for processing YAML data (see methods `parseYAML`